+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v1.01](https://github.com/realityforge/keycloak-jaxrs-client-authfilter/tree/v1.01) (2019-08-30)
 [Full Changelog](https://github.com/realityforge/keycloak-jaxrs-client-authfilter/compare/v1.00...v1.01)