+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v1.02](https://github.com/realityforge/keycloak-jaxrs-client-authfilter/tree/v1.02) (2019-09-09)
 [Full Changelog](https://github.com/realityforge/keycloak-jaxrs-client-authfilter/compare/v1.01...v1.02)